pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ran-llm-xapp"
version = "0.1.0"
description = "Pure Python synthetic reproduction of LLM-xApp RRM metrics and curves"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "matplotlib>=3.7",
  "PyYAML>=6.0",
]

[tool.setuptools.packages.find]
include = ["ran_llm_xapp*", "scripts*"]
